requirements.txt
python-telegram-bot==20.5
yt-dlp==2023.09.24
python-dotenv==1.0.0